+++ include/xkbcommon/xkbcommon.h
@@ -3261,6 +3261,12 @@ xkb_machine_builder_update_a11y(
 );
 
 /**
+ * @struct xkb_machine_builder_mods_remap_update
+ * @ingroup abi-struct-contract
+ *
+ * Modifiers remapping update for
+ * `xkb_machine_builder::xkb_machine_builder_update_mods_remap()`
+ *
  * Remap a modifier combination, e.g. to make `Control+Alt` act as
  * `LevelThree` (`AltGr`). This helps improve *compatibility* across platforms.
  *
@@ -3274,42 +3280,64 @@ xkb_machine_builder_update_a11y(
  * - There is no other remapping entry with the source modifiers being a
  *   superset of this entry. E.g. `Control+Alt` has priority over `Control`.
  *
+ * @figure@figcaption
+ * Example: remap `Control+Alt` to `LevelThree`.
+ * @endfigcaption
+ * @snippet "test/server-state.c" xkb_machine_builder_mods_remap_update_example
+ * @endfigure
+ *
+ * @sa `xkb_keymap::xkb_keymap_mod_get_mask()`
+ * @sa `xkb_machine_builder::xkb_machine_builder_update_mods_remap()`
+ *
+ * @since 1.14.0
+ */
+struct xkb_machine_builder_mods_remap_update {
+    /**
+     * Size of this structure in bytes.
+     *
+     * @sa @ref abi-struct-contract
+     *
+     * @since 1.14.0
+     */
+    uint32_t size;
+    /**
+     * Modifier combination to remap, using their [encoding].
+     * Must be non-zero, unless both #source and #target are 0 to clear all
+     * entries.
+     *
+     * @since 1.14.0
+     *
+     * [encoding]: @ref modifiers-encoding
+     */
+    xkb_mod_mask_t source;
+    /**
+     * Modifier combination to remap to, using their [encoding], or 0 to remove
+     * the entry for #source. If both #source and #target are 0, all
+     * entries are cleared.
+     *
+     * @since 1.14.0
+     *
+     * [encoding]: @ref modifiers-encoding
+     */
+    xkb_mod_mask_t target;
+};
+
+/**
+ * Update the modifiers remapping of an `xkb_machine_builder` object.
+ *
  * @param[in,out] builder The `xkb_machine` builder object to modify.
- * @param[in]     source  Modifier combination to remap, using their [encoding].
- *                        Must be non-zero, unless both @p source and @p target
- *                        are 0 to clear all entries.
- * @param[in]     target  Modifier combination to remap to, using their
- *                        [encoding], or 0 to remove the entry for @p source.
- *                        If both @p source and @p target are 0, all entries are
- *                        cleared.
+ * @param[in]     update  Modifiers remapping update object.
  *
  * @returns `::XKB_SUCCESS` on success, otherwise an error code.
- *
- * Example:
- *
- * ```c
- * struct xkb_keymap *keymap = xkb_machine_builder_get_keymap(builder);
- * // Remap Control+Alt to LevelThree (AltGr)
- * const xkb_mod_mask_t ctrl = xkb_keymap_mod_get_mask(keymap, XKB_MOD_NAME_CTRL);
- * const xkb_mod_mask_t alt = xkb_keymap_mod_get_mask(keymap, XKB_VMOD_NAME_ALT);
- * const xkb_mod_mask_t level3 = xkb_keymap_mod_get_mask(keymap, XKB_VMOD_NAME_LEVEL3);
- * if (xkb_machine_builder_remap_mods(builder, ctrl | alt, level3)) {
- *     // handle error
- *     …
- * }
- * ```
  *
  * @since 1.14.0
  *
  * @memberof xkb_machine_builder
- *
- * [encoding]: @ref modifiers-encoding
  */
 XKB_EXPORT enum xkb_error_code
-xkb_machine_builder_remap_mods(
+xkb_machine_builder_update_mods_remap(
     struct xkb_machine_builder *builder,
-    xkb_mod_mask_t source,
-    xkb_mod_mask_t target
+    const struct xkb_machine_builder_mods_remap_update *update
 );
 
 /**